Removed c lpm code,change dev_x_dot into sys_vvX_maps and made freq calc use pwm data.

+++ pulsefire-java/pulsefire-device-ui/src/main/java/org/nongnu/pulsefire/device/ui/tabs/JTabPanelPwm.java
@@ -348,7 +348,7 @@ public class JTabPanelPwm extends AbstractFireTabPanel implements DeviceCommandL
 		checkChannels(steps);
 	}
 	
-	private void checkChannels(int steps) {
+	private void checkChannels(long steps) {
 		Boolean limit = PulseFireUI.getInstance().getSettingsManager().getSettingBoolean(PulseFireUISettingKeys.LIMIT_CHANNELS);
 		if (limit==false) {
 			steps = CommandName.pulse_steps.getMaxValue();
